T-GCN/T-GCN-PyTorch/requirements.txt
numpy
matplotlib
pandas
torch
pytorch-lightning>=1.3.0
torchmetrics>=0.3.0
python-dotenv
